# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMapX API",
    "description": "Generated from the API's own route table and the integration route registry. Do not edit by hand — run `pnpm openapi:generate`.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/saved/labels/{label}": {
      "put": {
        "operationId": "putApiSavedLabelsLabel",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "label",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "lat",
                  "lng"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "placeId": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      },
      "delete": {
        "operationId": "deleteApiSavedLabelsLabel",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "label",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
    "/api/saved/lists": {
      "get": {
        "operationId": "getApiSavedLists",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      },
      "post": {
        "operationId": "postApiSavedLists",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "isPrivate": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
    "/api/saved/lists/{id}": {
      "delete": {
        "operationId": "deleteApiSavedListsId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      },
      "patch": {
        "operationId": "patchApiSavedListsId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "isPrivate": {
                    "type": "boolean"
                  },
                  "sortOrder": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
    "/api/saved/lists/{id}/export": {
      "get": {
        "operationId": "getApiSavedListsIdExport",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "gpx",
                "geojson",
                "kml"
              ],
              "default": "geojson"
            },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
    "/api/saved/lists/{id}/places": {
      "get": {
        "operationId": "getApiSavedListsIdPlaces",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      },
      "post": {
        "operationId": "postApiSavedListsIdPlaces",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "lat",
                  "lng"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "placeId": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
    "/api/saved/places/{id}": {
      "delete": {
        "operationId": "deleteApiSavedPlacesId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      },
      "patch": {
        "operationId": "patchApiSavedPlacesId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "sortOrder": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "saved"
        ],
        "x-openmapx-auth": "session"
      }
    },
...
  }
}